# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.SignInRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.SignInResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
      user:
        $ref: '#/definitions/internals_user_controller_dto.User'
    required:
    - accessToken
    - refreshToken
    - user
    type: object
  dto.SignUpResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
      user:
        $ref: '#/definitions/internals_user_controller_dto.User'
    required:
    - accessToken
    - refreshToken
    - user
    type: object
...
  internals_user_controller_dto.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      updated_at:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
    type: object
host: localhost:8080
info:
  contact:
    email: anhquoc18092003@gmail.com
    name: Tran Phuoc Anh Quoc
  description: Swagger API for Go Clean Architecture.
  license:
    name: MIT
    url: https://github.com/MartinHeinz/go-project-blueprint/blob/master/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Ecommerce Clean Architecture Swagger API
  version: "1.0"
paths:
  /auth/signin:
    post:
      consumes:
      - application/json
      description: Authenticates the user based on the provided credentials and returns
        access tokens and user info if successful.
      parameters:
      - description: User sign-in request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully signed in
          schema:
            $ref: '#/definitions/dto.SignInResponse'
        "400":
          description: Bad Request - Invalid parameters
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict - Wrong password or Email does not exist
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error - Failed to sign in
          schema:
            $ref: '#/definitions/response.Response'
      summary: User Sign-In
      tags:
      - Auth
  /auth/signout:
    post:
      description: Logs out the authenticated user by invalidating the current session
        token.
      produces:
      - application/json
      responses:
        "200":
          description: User successfully logged out
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request - Missing Authorization header
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized - Invalid or missing user ID
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error - Failed to sign out
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: User Sign-Out
      tags:
      - Auth
  /auth/signup:
    post:
      consumes:
      - multipart/form-data
      description: Registers a new user with the provided details and returns access
        tokens along with user info if successful.
      parameters:
      - description: User email (must be unique)
        in: formData
        name: email
        required: true
        type: string
      - description: User name (must be unique)
        in: formData
        name: name
        required: true
        type: string
      - description: User password
        in: formData
        name: password
        required: true
        type: string
      - description: User avatar file
        in: formData
        name: avatar
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: User successfully registered
          schema:
            $ref: '#/definitions/dto.SignUpResponse'
        "400":
          description: Bad Request - Invalid parameters
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict - Email or Name already in use
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error - Failed to sign up
          schema:
            $ref: '#/definitions/response.Response'
      summary: User Sign-Up
      tags:
      - Auth
...
  /users:
    get:
      description: Retrieves a paginated list of users based on search criteria.
      parameters:
      - description: Search keyword for filtering users
        in: query
        name: search
        type: string
      - description: Page number for pagination
        in: query
        name: page
        type: integer
      - description: Number of users per page
        in: query
        name: size
        type: integer
      - description: Column name to sort by
        in: query
        name: order_by
        type: string
      - description: Sort in descending order (true/false)
        in: query
        name: order_desc
        type: boolean
      - description: Retrieve all users without pagination
        in: query
        name: take_all
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved users list
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request - Invalid query parameters
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error - Failed to get users
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Get Users List
      tags:
      - Users
  /users/{id}:
    delete:
      description: Deletes a user from the database by ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User successfully deleted
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request - Invalid user ID format
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found - User does not exist
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error - Failed to delete user
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete User
      tags:
      - Users
    get:
      description: Retrieves detailed information of a specific user by ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved user details
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request - Invalid user ID format
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found - User does not exist
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error - Failed to retrieve user details
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Get User Detail
      tags:
      - Users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"